Adapted version of autopopulate.py implements custom arguments for make()
Additional custom keyword arguments can be given to the populate() function, which will be passed down directly to each make() call of that table.
That table has to expect the exact name of the custom argument (see common_img.Segmentation for an example).

This script version has to replace the installed DataJoint version of autopopulate.py before running pipeline!

Signed-off-by: Hendriela <hendrik-heiser@gmx.de>